+++ lib/rune.dart
@@ -25,6 +25,8 @@ export 'src/registry/component_registry.dart' show ComponentRegistry;
 export 'src/registry/constant_registry.dart' show ConstantRegistry;
 export 'src/registry/extension_registry.dart'
     show ExtensionRegistry, RuneExtensionHandler;
+export 'src/registry/imperative_registry.dart'
+    show ImperativeHandler, ImperativeRegistry;
 export 'src/registry/registry.dart' show Registry;
 export 'src/registry/value_registry.dart' show ValueRegistry;
 export 'src/registry/widget_registry.dart' show WidgetRegistry;